# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Service Booking API",
        "description": "A Laravel-based API for booking services.",
        "contact": {
            "email": "smazoomder@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/bookings": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create a new service booking",
                "description": "This endpoint allows customers to book a service by providing their name, phone number, and service ID.",
                "operationId": "storeBooking",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "service_id",
                                    "name",
                                    "phone",
                                    "schedule_date"
                                ],
                                "properties": {
                                    "service_id": {
                                        "description": "Service UUID",
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "name": {
                                        "description": "Customer name",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Customer phone number",
                                        "type": "string"
                                    },
                                    "schedule_date": {
                                        "description": "Scheduled date and time for service",
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "service_id": "028c9fb6-ebfd-4003-ad5c-681ff37263d5",
                                "name": "Shaon",
                                "phone": "+8801832540116",
                                "schedule_date": "2025-05-09T12:51:22.653Z"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "booking_id": {
                                            "description": "Booking ID",
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "status": {
                                            "description": "Booking status",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request data"
                    },
                    "404": {
                        "description": "Service not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/v1/bookings/{bookingId}": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Retrieve the status of a booking",
                "description": "This endpoint allows customers to check the status of their booking using a unique booking ID.",
                "operationId": "getBookingStatus",
                "parameters": [
                    {
                        "name": "bookingId",
                        "in": "path",
                        "description": "UUID of the booking",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking status retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Booking retrieved from cache."
                                        },
                                        "data": {
                                            "properties": {
                                                "booking_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "e183b1a8-c0b3-41a2-be77-4b16a9670c01"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "service": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "format": "uuid",
                                                            "example": "028c9fb6-ebfd-4003-ad5c-681ff37263d5"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Cleaning Service"
                                                        },
                                                        "category": {
                                                            "type": "string",
                                                            "example": "Cleaning"
                                                        },
                                                        "price": {
                                                            "type": "string",
                                                            "example": "100.00"
                                                        },
                                                        "description": {
                                                            "type": "string",
                                                            "example": "Cleaning description goes here."
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "schedule_date_time": {
                                                    "type": "string",
                                                    "format": "datetime",
                                                    "example": "2025-05-09 12:51:22"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Booking not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a valid bearer token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}